ostconfin/requirements.txt
playwright==1.44.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
lxml==5.2.2
